Removed instruction order test
